feat: implement routing configuration and initialize application rendering
