+++ preact/src/shared/utils/routing/index.ts
@@ -1,3 +1,1 @@
-export * from '../../hooks/use-router';
-export * from './routes';
 export * from './routes-config';
